(NomenCleaner) 3rd version has been integrated into HATK.
