+++ src/HLA_Study.py
@@ -328,13 +328,11 @@ class HLA_Study(object):
                 ### NomenCleaner
                 from NomenCleaner.NomenCleaner import HATK_NomenCleaner
 
-                myNomenCleaner = HATK_NomenCleaner(_args.iat, _args.imgt, _args.out,
-                                                   _args.hped, _args.hped_G, _args.hped_P,
-                                                   _args.oneF, _args.twoF, _args.threeF, _args.fourF,
-                                                   _args.G_group, _args.P_group, _args.old_format,
+                myNomenCleaner = HATK_NomenCleaner(_args.hat, _args.imgt, _args.out, _args.hped,
+                                                   _args.oneF, _args.twoF, _args.threeF, _args.fourF, _args.Ggroup, _args.Pgroup,
                                                    __f_NoCaption=_args.NoCaption, __leave_NotFound=_args.leave_NotFound)
 
-                print(std_MAIN_PROCESS_NAME + "NomenCleaner result : \n{}".format(myNomenCleaner.getResults()))
+                print(std_MAIN_PROCESS_NAME + "NomenCleaner result : \n{}".format(myNomenCleaner.chped))
 
             elif _args.logistic:
